Int is not enough

+++ tracestruct.h
@@ -1,6 +1,6 @@
 
 typedef struct trace_record {
-  int address;
+  long int address;
   char size;
   bool is_write;
 } mtrace_t;
